Modify find command to find by category and product preference
New predicate classes are added to test for category (a.k.a tags)
and product preference of the clients. However, the client should
only be used to test with ClientSatisfyAllPredicate.

This is due to the constraint of only being able to pass a single
predicate class into the find command to test for.

The product preference predicate class should be updated to test
for the correct client attribute after the client class is
updated accordingly.