Add template for comparator and change UI to accept a sorted list
A template for comparator is added for comparing specified
attributes in the Client class.

The UI that dynamically updates the client list accepts a
sorted version of the filtered list.

This is so that the UI is able to reflect any changes made to the
client list by sorting and filtering at the same time.